requirements.txt
asgiref==3.8.1
django-bootstrap5==24.3
django==5.1.1
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
